+++ Assets/Scripts/GameManager/GameManager.cs
@@ -9,11 +9,11 @@ public class GameManager : MonoBehaviour
     public enum CurrentPlayer { Poule, Lapin, None }// pour pas que le joueur joue les deux perso
     public CurrentPlayer WhoIsPlayer;
 
-    public enum GameStat { ReStart, Draw, Wait, Fight, Result }// Etat actuel du combat -> LE MEME POUR TOUT LES JOUEURS
+    public enum GameStat { ReStart, Draw, Wait, Fight, Result }// Etat actuel du combat -> LE MEME POUR TOUT LES JOUEURS 
     public GameStat CurrentStat;
 
-    public int PouleAction;
-    public int LapinAction;
+    public int PouleAction = -1;
+    public int LapinAction = -1;
 
     public GameObject PlayerPoule;
     public GameObject PlayerPouleGrid;
@@ -32,6 +32,7 @@ public class GameManager : MonoBehaviour
     public TextMeshProUGUI PaTXT;
 
     public Animator animPoule;
+    public Animator animLapin;
 
     public int[,] T_Drawings = {
         //Attaques
@@ -56,9 +57,9 @@ public class GameManager : MonoBehaviour
         {3, 4, 5 }          //Soin horizontal
     };
 
-    //Boolï¿½en ï¿½ valider une fois la rï¿½ception d'une attaque faite, et ï¿½ invalider une fois la phase passï¿½e.
+    //Booléen à valider une fois la réception d'une attaque faite, et à invalider une fois la phase passée.
     private bool bReceptionReseau = false;
-    //Rï¿½fï¿½rence au manager rï¿½seau
+    //Référence au manager réseau
     SteamManagerGame steamManagerGame=null;
 
     // Start is called before the first frame update
@@ -75,11 +76,11 @@ public class GameManager : MonoBehaviour
             WhoIsPlayer = CurrentPlayer.Poule;
         }
 
-        //Mise en place du manager de rï¿½seau.
+        //Mise en place du manager de réseau.
         steamManagerGame = GameObject.Find("SteamManagerGame").GetComponent<SteamManagerGame>();
         if (steamManagerGame == null)
         {
-            Debug.Log("Erreur lors de la rï¿½cupï¿½ration du manager rï¿½seau.");
+            Debug.Log("Erreur lors de la récupération du manager réseau.");
         }
 
         CurrentStat = GameStat.Draw;
@@ -97,10 +98,15 @@ public class GameManager : MonoBehaviour
                 PlayerPoule.GetComponent<PlayerControl>().ActionPoint = 3;
                 PlayerLapin.GetComponent<PlayerControl>().ActionPoint = 3;
                 ResultGrid.SetActive(false);
-                animPoule.SetBool("Attaque1", false);
-                animPoule.SetBool("Attaque2", false);
-                animPoule.SetBool("Defense1", false);
-                animPoule.SetBool("Defense2", false);
+                animPoule.SetBool("Attaque", false);
+                animPoule.SetBool("Esquive", false);
+                animPoule.SetBool("Parade", false);
+                animPoule.SetBool("Soin", false);
+                animLapin.SetBool("Attaque", false);
+                animLapin.SetBool("Esquive", false);
+                animLapin.SetBool("Parade", false);
+                animLapin.SetBool("Soin", false);
+
 
                 for (i = 0; i < T_PlayerPouleGrid.Length; i++)
                 {
@@ -108,17 +114,15 @@ public class GameManager : MonoBehaviour
                     if(T_PlayerPouleGrid[i].color == Color.white) {
                         T_PlayerPouleGrid[i].color = Color.white;
                     } else {
-                        T_PlayerPouleGrid[i].color = Color.red;
+                        T_PlayerPouleGrid[i].color = Color.gray;
                     }
 
                     if (T_PlayerLapinGrid[i].color == Color.white) {
                         T_PlayerLapinGrid[i].color = Color.white;
                     } else {
-                        T_PlayerLapinGrid[i].color = Color.red;
+                        T_PlayerLapinGrid[i].color = Color.gray;
                     }
-
-                    WhoIsPlayer = CurrentPlayer.Poule;
-
+                    
                     //Debug.Log("Couleur apres : " + T_PlayerPouleGrid[i].color);
                 }
                 break;
@@ -126,34 +130,32 @@ public class GameManager : MonoBehaviour
             case GameStat.Draw:
                 drawPhase();
 
-                if (WhoIsPlayer == CurrentPlayer.Poule)
-                {
-                    WhoIsPlayerTxT.SetText("Vous ï¿½tes : Poule");
+                if (WhoIsPlayer == CurrentPlayer.Poule) {
+                    WhoIsPlayerTxT.SetText("Vous êtes : Poule");
                     PaTXT.SetText("Point d'action : " + PlayerPoule.GetComponent<PlayerControl>().ActionPoint);
                     PlayerLapinGrid.SetActive(false);
                     PlayerPouleGrid.SetActive(true);
-                }
-                else if (WhoIsPlayer == CurrentPlayer.Lapin)
-                {
-                    WhoIsPlayerTxT.SetText("Vous ï¿½tes : Lapin");
+
+                } else if (WhoIsPlayer == CurrentPlayer.Lapin) {
+                    WhoIsPlayerTxT.SetText("Vous êtes : Lapin");
                     PaTXT.SetText("Point d'action : " + PlayerLapin.GetComponent<PlayerControl>().ActionPoint);
                     PlayerLapinGrid.SetActive(true);
                     PlayerPouleGrid.SetActive(false);
+
                 }
                 break;
             case GameStat.Wait:
-                //Attente de la rï¿½ception d'une rï¿½ponse rï¿½seau.
+                //Attente de la réception d'une réponse réseau.
                 if (bReceptionReseau==true)
                 {
                     CurrentStat = GameStat.Fight;
+                    bReceptionReseau = false;
                 }
                 break;
             case GameStat.Fight:
                 resultPhase();
-
                 CurrentStat = GameStat.Result;
                 bReceptionReseau = false;
-
                 break;
         }
 
@@ -174,17 +176,18 @@ public class GameManager : MonoBehaviour
             // POUR LA POULE
             if (WhoIsPlayer == CurrentPlayer.Poule)
             {
-                playerDrawing(PlayerPoule, T_PlayerPouleGrid, out T_PouleDrawing);
+                PouleAction = playerDrawing(PlayerPoule, T_PlayerPouleGrid, out T_PouleDrawing);
+
             }
-            //POUR LE LAPIN
+            //POUR LE LAPIN 
             if (WhoIsPlayer == CurrentPlayer.Lapin)
             {
-                playerDrawing(PlayerLapin, T_PlayerLapinGrid, out   T_LapinDrawing);
+                LapinAction = playerDrawing(PlayerLapin, T_PlayerLapinGrid, out   T_LapinDrawing);
             }
         }
 
 
-        //Changement d'ï¿½tat quand le joueur en cours n'a plus de PA
+        //Changement d'état quand le joueur en cours n'a plus de PA
         /*if (WhoIsPlayer == CurrentPlayer.Lapin && (PlayerLapin.GetComponent<PlayerControl>().ActionPoint == 0))
         {
 
@@ -197,17 +200,18 @@ public class GameManager : MonoBehaviour
             CurrentStat = GameStat.Wait;
         }*/
 
-        //Changement d'ï¿½tat quand les deux joueurs on plus de PA
+        //Changement d'état quand les deux joueurs on plus de PA
         /*if ((PlayerPoule.GetComponent<PlayerControl>().ActionPoint == 0) && (PlayerLapin.GetComponent<PlayerControl>().ActionPoint == 0))
         {
             CurrentStat = GameStat.Fight;
         }*/
     }
 
-    void playerDrawing(GameObject PlayerObject, SpriteRenderer[] T_PlayerGrid, out int[] T_PlayerDrawing)
+    int playerDrawing(GameObject PlayerObject, SpriteRenderer[] T_PlayerGrid, out int[] T_PlayerDrawing)
     {
         int i;
-        int numDrawing, typeDrawing;
+        int numDrawing;
+        int typeDrawing = -1;
         int j = 0;
         int[] T_BaseDrawing = { -1, -1, -1 };
         T_PlayerDrawing = T_BaseDrawing;
@@ -245,21 +249,21 @@ public class GameManager : MonoBehaviour
                         if (numDrawing != -1)
                         {
                             typeDrawing = (numDrawing / 4) + 1;
-                            Debug.Log("Dessin nï¿½" + numDrawing);
+                            Debug.Log("Dessin n°" + numDrawing);
                             PlayerAttack = new AttaqueV3(typeDrawing, T_PlayerDrawing[0], T_PlayerDrawing[1], T_PlayerDrawing[2]);
                             //WhoIsPlayer = CurrentPlayer.Lapin;
 
 
-                            //Envoie de l'action au rï¿½seau.
+                            //Envoie de l'action au réseau.
                             if (steamManagerGame.sendActionToSteamNetwork(PlayerAttack))
                             {
-                                Debug.Log("Rï¿½ussite de l'envoie de l'attaque rï¿½seau.");
-                                //Passage en phase d'attente de rï¿½ponse rï¿½seau.
+                                Debug.Log("Réussite de l'envoie de l'attaque réseau.");
+                                //Passage en phase d'attente de réponse réseau.
                                 CurrentStat = GameStat.Wait;
                             }
                             else
                             {
-                                Debug.Log("ï¿½chec de l'envoie de l'attaque rï¿½seau.");
+                                Debug.Log("Échec de l'envoi de l'attaque réseau.");
                             }
                         }
                         else
@@ -267,7 +271,7 @@ public class GameManager : MonoBehaviour
                             Debug.Log("Dessin invalide");
                             PlayerObject.GetComponent<PlayerControl>().ActionPoint = 3;
                             for(i=0; i<T_PlayerGrid.Length; i++) {
-                                if (T_PlayerGrid[i].color != Color.red) {
+                                if (T_PlayerGrid[i].color != Color.gray) {
                                     T_PlayerGrid[i].color = Color.white;
                                 }
                             }
@@ -283,16 +287,18 @@ public class GameManager : MonoBehaviour
                 }
                 else
                 {
-                    Debug.Log("Plus de PA ! ");
+                    Debug.Log("Cette case ne peut pas etre cochée !");
                 }
             }
         }
+
+        return typeDrawing;
     }
 
     void resultPhase()
     {
         int i;
-
+        
         for(i=0; i<T_ResultGrid.Length; i++) {
             T_ResultGrid[i].color = Color.white;
         }
@@ -312,10 +318,10 @@ public class GameManager : MonoBehaviour
                 }
             }
 
-            if(T_PlayerPouleGrid[i].color == Color.red) {
+            if(T_PlayerPouleGrid[i].color == Color.gray) {
                 T_PlayerPouleGrid[i].color = Color.white;
             }
-            if (T_PlayerLapinGrid[i].color == Color.red)
+            if (T_PlayerLapinGrid[i].color == Color.gray)
             {
                 T_PlayerLapinGrid[i].color = Color.white;
             }
@@ -326,9 +332,9 @@ public class GameManager : MonoBehaviour
         PlayerPouleGrid.SetActive(false);
         ResultGrid.SetActive(true);
 
-
+        
         /*
-        //Pour la poule
+        //Pour la poule 
         //Attaque :
         if (T_PlayerPouleGrid[0].color == Color.black && T_PlayerPouleGrid[1].color == Color.black && T_PlayerPouleGrid[2].color == Color.black)
         {
@@ -342,7 +348,7 @@ public class GameManager : MonoBehaviour
             animPoule.SetBool("Attaque2", true);
 
         }
-        //Defense :
+        //Defense : 
         else if (T_PlayerPouleGrid[8].color == Color.black && T_PlayerPouleGrid[7].color == Color.black && T_PlayerPouleGrid[6].color == Color.black)
         {
             Debug.Log("DEFENSE 1 DE LA POULE !");
@@ -357,7 +363,7 @@ public class GameManager : MonoBehaviour
             animPoule.SetBool("Defense2", true);
         }
 
-        //Pour le Lapin
+        //Pour le Lapin 
         //Attaque :
         if (T_PlayerLapinGrid[0].color == Color.black && T_PlayerLapinGrid[1].color == Color.black && T_PlayerLapinGrid[2].color == Color.black)
         {
@@ -370,7 +376,7 @@ public class GameManager : MonoBehaviour
             Debug.Log("ATTAQUE 2 DU LAPIN !");
             LapinAction = Action.Attaque2;
         }
-        //Defense :
+        //Defense : 
         else if (T_PlayerLapinGrid[8].color == Color.black && T_PlayerLapinGrid[7].color == Color.black && T_PlayerLapinGrid[6].color == Color.black)
         {
             Debug.Log("DEFENSE 1 DU LAPINE !");
@@ -407,10 +413,13 @@ public class GameManager : MonoBehaviour
             }
         }
 
+        Debug.Log("Nombre de collisons : " + nbCollide);
+
         //ACTION DE DEFENSE
         if (PouleAction == 2 ^ LapinAction == 2) {
             if (PouleAction == 2) {
-                animPoule.SetBool("Defense1", true);
+                Debug.Log("Poule: Defense");
+                animPoule.SetBool("Parade", true);
                 if (nbCollide == 2) {
                     LapinAction = 0;
                     tapeLapin = true;
@@ -418,6 +427,8 @@ public class GameManager : MonoBehaviour
                     PouleAction = 0;
                 }
             } else {
+                Debug.Log("Lapin: Defense");
+                animLapin.SetBool("Parade", true);
                 if (nbCollide == 2) {
                     PouleAction = 0;
                     tapePoule = true;
@@ -427,7 +438,7 @@ public class GameManager : MonoBehaviour
             }
 
             nbCollide = 0;
-
+            
         }
 
 
@@ -435,7 +446,8 @@ public class GameManager : MonoBehaviour
         //ACTION D'ESQUIVE
         if(PouleAction == 3 || LapinAction == 3) {
             if(PouleAction == 3) {
-                animPoule.SetBool("Attaque2", true);
+                Debug.Log("Poule: Esquive");
+                animPoule.SetBool("Esquive", true);
                 if (nbCollide == 1) {
                     LapinAction = 0;
                 }
@@ -444,6 +456,8 @@ public class GameManager : MonoBehaviour
                 }
             }
             if(LapinAction == 3) {
+                Debug.Log("Lapin: Esquive");
+                animLapin.SetBool("Esquive", true);
                 if (nbCollide == 1) {
                     PouleAction = 0;
                 }
@@ -462,13 +476,16 @@ public class GameManager : MonoBehaviour
         //ACTION D'ATTAQUE
         if(PouleAction == 1 || LapinAction == 1) {
             if(PouleAction == 1) {
-                animPoule.SetBool("Attaque1", true);
-                if(nbCollide == 0) {
+                Debug.Log("Poule: Attaque");
+                animPoule.SetBool("Attaque", true);
+                if(nbCollide == 0) { 
                     tapeLapin = true;
                 }
             }
             if(LapinAction == 1) {
-                if(nbCollide == 0) {
+                Debug.Log("Lapin: Attaque");
+                animLapin.SetBool("Attaque", true);
+                if (nbCollide == 0) { 
                     tapePoule = true;
                 }
             }
@@ -479,13 +496,16 @@ public class GameManager : MonoBehaviour
         //ACTION DE SOIN
         if(PouleAction == 4 || LapinAction == 4) {
             if(PouleAction == 1) {
-                animPoule.SetBool("Defense2", true);
-                if(nbCollide < 2) {
+                Debug.Log("Poule: Soin");
+                animPoule.SetBool("Soin", true);
+                if(nbCollide < 2) { 
                     healPoule = true;
                 }
             }
             if(LapinAction == 1) {
-                if(nbCollide < 2) {
+                Debug.Log("Lapin: Soin");
+                animLapin.SetBool("Soin", true);
+                if (nbCollide < 2) { 
                     healLapin = true;
                 }
             }
@@ -513,26 +533,30 @@ public class GameManager : MonoBehaviour
         }
 
 
-
+        
 
 
         //yield on a new YieldInstruction that waits for 5 seconds.
         yield return new WaitForSeconds(2);
-
+        
         if(tapeLapin) {
-            PlayerLapin.GetComponent<PlayerControl>().Life--;
+            Debug.Log("Degats sur le lapin");
+            PlayerLapin.GetComponent<PlayerControl>().ThisPlayerGetHit = true;
         }
         if(tapePoule) {
-            PlayerPoule.GetComponent<PlayerControl>().Life--;
+            Debug.Log("Degats sur la poule");
+            PlayerPoule.GetComponent<PlayerControl>().ThisPlayerGetHit = true;
         }
         if(healLapin) {
-            if(PlayerLapin.GetComponent<PlayerControl>().Life < 3) {
-                PlayerLapin.GetComponent<PlayerControl>().Life++;
+            Debug.Log("Soin sur le lapin");
+            if (PlayerLapin.GetComponent<PlayerControl>().Life < 3) {
+                PlayerLapin.GetComponent<PlayerControl>().ThisPlayerGetHeal = true;
             }
         }
         if(healPoule) {
+            Debug.Log("Soins sur le lapin");
             if(PlayerPoule.GetComponent<PlayerControl>().Life < 3) {
-                PlayerPoule.GetComponent<PlayerControl>().Life++;
+                PlayerPoule.GetComponent<PlayerControl>().ThisPlayerGetHeal = true;
             }
         }
 
@@ -547,15 +571,12 @@ public class GameManager : MonoBehaviour
         //After we have waited 5 seconds print the time again.
     }
 
-    int ValidDrawing(int[] T_PlayerDrawing)
-    {
+    int ValidDrawing(int[] T_PlayerDrawing) {
         int i;
         int valid = -1;
 
-        for (i = 0; i < 16; i++)
-        {
-            if (T_PlayerDrawing[0] == T_Drawings[i, 0] && T_PlayerDrawing[1] == T_Drawings[i, 1] && T_PlayerDrawing[2] == T_Drawings[i, 2])
-            {
+        for (i = 0; i < 16; i++) {
+            if (T_PlayerDrawing[0] == T_Drawings[i, 0] && T_PlayerDrawing[1] == T_Drawings[i, 1] && T_PlayerDrawing[2] == T_Drawings[i, 2]) {
                 valid = i;
                 break;
             }
@@ -564,32 +585,43 @@ public class GameManager : MonoBehaviour
         return valid;
     }
 
-    //BUT : Rï¿½cupï¿½rer l'attaque en rï¿½seau et signaler que l'attaque est reï¿½ue.
+    //BUT : Récupérer l'attaque en réseau et signaler que l'attaque est reçue.
     public void ReceptionAttaqueReseau(AttaqueV3 attaqueV3)
     {
-        if (bReceptionReseau==false)
-        {
-            //ICI GAETAN UTILISER L'ATTAQUE V3 POUR METTRE EN PLACE L'ATTAQUE DE L'ADVERSAIRE.
-            if (WhoIsPlayer == CurrentPlayer.Poule)
-            {
-                //Le joueur est la poule, donc l'attaque reï¿½ue en rï¿½seau s'applique au lapin
+        int i;
+
+        if (bReceptionReseau==false) {
+            if (WhoIsPlayer == CurrentPlayer.Poule) {
+                //Le joueur est la poule, donc l'attaque reçue en réseau s'applique au lapin
+                Debug.Log("Reception de l'attaque du lapin");
                 LapinAction = attaqueV3.Type;
                 T_LapinDrawing[0] = attaqueV3.Case1;
                 T_LapinDrawing[1] = attaqueV3.Case2;
                 T_LapinDrawing[2] = attaqueV3.Case3;
 
-                Debug.Log("ï¿½criture de l'attaque de Lapin.");
+                for(i=0; i<T_PlayerLapinGrid.Length; i++) {
+                    if(i == T_LapinDrawing[0] || i == T_LapinDrawing[1] || i == T_LapinDrawing[2]) {
+                        T_PlayerLapinGrid[i].color = Color.black;
+                    } else {
+                        T_PlayerLapinGrid[i].color = Color.white;
+                    }
+                }
 
-            }
-            else if (WhoIsPlayer == CurrentPlayer.Lapin)
-            {
-                //Le joueur est le lapin, donc l'attaque reï¿½ue en rï¿½seau s'applique ï¿½ la poule.
+            } else if (WhoIsPlayer == CurrentPlayer.Lapin) {
+                //Le joueur est le lapin, donc l'attaque reçue en réseau s'applique à la poule.
+                Debug.Log("Reception de l'attaque de la poule");
                 PouleAction = attaqueV3.Type;
                 T_PouleDrawing[0] = attaqueV3.Case1;
                 T_PouleDrawing[1] = attaqueV3.Case2;
                 T_PouleDrawing[2] = attaqueV3.Case3;
 
-                Debug.Log("ï¿½criture de l'attaque de Poule.");
+                for(i=0; i<T_PlayerPouleGrid.Length; i++) {
+                    if(i == T_PouleDrawing[0] || i == T_PouleDrawing[1] || i == T_PouleDrawing[2]) {
+                        T_PlayerPouleGrid[i].color = Color.black;
+                    } else {
+                        T_PlayerPouleGrid[i].color = Color.white;
+                    }
+                }
 
             }
 
@@ -599,10 +631,10 @@ public class GameManager : MonoBehaviour
 
 
 
-    //BUT : Mettre fin au rï¿½seau, au jeu, et rediriger ï¿½ l'ï¿½cran titre.
+    //BUT : Mettre fin au réseau, au jeu, et rediriger à l'écran titre.
     void EndGame()
     {
-        //Rï¿½fï¿½rence au manager rï¿½seau.
+        //Référence au manager réseau.
         steamManagerGame.endLobby();
         //Redirection de fin de jeu.
         SceneManager.LoadScene("TitleScreen");