+++ changelog.md
@@ -1,3 +1,17 @@
+## [0.6.3] - 2022-09-03 - G.P.
+###Modification
+-Correction du bug empechant la lecture des collisions entre cases de joueurs
+
+## [0.6.2] - 2022-09-03 - G.P.
+###Modification
+-Correction du bug de l'animator du Lapin
+
+## [0.6.1] - 2022-09-03 - G.P.
+###Ajout
+-Animator du Lapin
+###Modification
+-Correction du bug empechant la lecture de l'action du joueur local
+
 ## [0.5.2] - 2022-08-03 - G.P.
 ###Ajout
 -Systeme de verification des dessins